experiments/robot/aloha/requirements_aloha.txt
numpy<2
draccus
torchvision
torch
pyquaternion
pyyaml
rospkg
pexpect
mujoco==2.3.7
dm_control==1.0.14
opencv-python
matplotlib
einops
packaging
h5py
traitlets
ipdb
IPython
modern_robotics
Pillow
termcolor
imageio[ffmpeg]
uvicorn
fastapi
requests
json_numpy
